third_party/WebKit/LayoutTests/fast/scrolling/absolute-position-behind-scrollbar.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
#a {
  position: absolute;
  background-color: #def;
  left: 100px;
  top: 100px;
  width: 1000px;
  height: 1000px;
}
</style>
Tests that the scrollbar can be clicked even when it clips absolute-positioned content.
<div id="a"></div>
<script>

eventSender.mouseMoveTo(790, 500);
eventSender.mouseDown();
eventSender.mouseUp();

shouldBeTrue(String(scrollY > 0));

</script>
