third_party/WebKit/LayoutTests/fast/scrolling/absolute-position-behind-scrollbar-expected.txt

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE
Tests that the scrollbar can be clicked even when it clips absolute-positioned content.
